+++ plugin-maven/CHANGES.md
@@ -6,6 +6,10 @@ We adhere to the [keepachangelog](https://keepachangelog.com/en/1.0.0/) format (
 ### Changes
 - Bump default `adocfmt` version `0.2.0` -> `0.3.1`, which adds table formatting support (`<formatTables>`, `<tableLayout>`, `<tableMaxLineWidth>`, `<tableBlankLines>`).
 
+### Changes
+- Add support to apply alternate license header within same format ([#872](https://github.com/diffplug/spotless/issues/872))
+- Add support to skip license header application based on source file content pattern ([#650](https://github.com/diffplug/spotless/issues/650)).
+
 ## [3.9.0] - 2026-07-27
 ### Added
 - Add support for Java formatting via [`prince-of-space`](https://github.com/agustafson/prince-of-space) with the new `<princeOfSpace>` step. ([#2991](https://github.com/diffplug/spotless/pull/2991))